+++ packages/target-tauri/src-tauri/src/state/mod.rs
@@ -2,3 +2,4 @@ pub(crate) mod app;
 pub(crate) mod deltachat;
 pub(crate) mod html_email_instances;
 pub(crate) mod menu_manager;
+pub(crate) mod translations;